navanrfp/requirements.txt
fastapi
uvicorn
requests
urllib3
pillow
pyzbar
python-multipart